// packages/cli-beacon/src/supported-origins/forge-v1/schemas.ts
export const FORGE_COMPILER_OUTPUT_WITH_BUILD_INFO_OPTION_FORMAT =
  "ethers-rs-sol-build-info-1";

export const ForgeCompilerOutputWithBuildInfoOptionSchema = z.object({
  id: z.string(),
  // Mapping from contract ID as number (e.g. "0", "1", etc.) to the source file path
  // This is needed to resolve the source files when the output JSON doesn't include the source file paths.
  source_id_to_path: z.record(z.string(), z.string()),
  language: z.enum(["Solidity", "Yul", "SolidityAST", "EVMAssembly"]),
  _format: z.literal(FORGE_COMPILER_OUTPUT_WITH_BUILD_INFO_OPTION_FORMAT),
  input: SolcJsonInputSchema.extend({
    // Repeat of the solc version
    version: z.string().optional(),
    // Additional paths
    allowPaths: z.array(z.string()).optional(),
    basePath: z.string().optional(),
    includePaths: z.array(z.string()).optional(),
  }),
  output: SolcJsonOutputSchema,
  solcLongVersion: z.string(),
  solcVersion: z.string().optional(),
});